Audio system now handles Guild Audio Region moves. AudioManager is now an interface so that non-public things could be implemented into AudioManager (AudioManagerImpl now) Added JDA#isAudioEnabled()
