+++ src/main/java/net/dv8tion/jda/audio/AudioConnection.java
@@ -26,10 +26,7 @@ import net.dv8tion.jda.utils.SimpleLog;
 import org.json.JSONObject;
 import tomp2p.opuswrapper.Opus;
 
-import java.net.DatagramPacket;
-import java.net.DatagramSocket;
-import java.net.NoRouteToHostException;
-import java.net.SocketException;
+import java.net.*;
 import java.nio.ByteBuffer;
 import java.nio.IntBuffer;
 import java.nio.ShortBuffer;
@@ -133,10 +130,12 @@ public class AudioConnection
         return channel.getGuild();
     }
 
-    public void close()
+    public void close(boolean regionChange)
     {
-        setSpeaking(false);
-        webSocket.close();
+//        setSpeaking(false);
+        sendThread.interrupt();
+        receiveThread.interrupt();
+        webSocket.close(regionChange);
     }
 
     private void setupSendThread()
@@ -149,7 +148,7 @@ public class AudioConnection
                 char seq = 0;           //Sequence of audio packets. Used to determine the order of the packets.
                 int timestamp = 0;      //Used to sync up our packets within the same timeframe of other people talking.
                 long lastFrameSent = System.currentTimeMillis();
-                while (!udpSocket.isClosed())
+                while (!udpSocket.isClosed() && !this.isInterrupted())
                 {
                     try
                     {
@@ -191,7 +190,15 @@ public class AudioConnection
                         LOG.warn("Closing AudioConnection due to inability to send audio packets.");
                         LOG.warn("Cannot send audio packet because JDA navigate the route to Discord.\n" +
                                 "Are you sure you have internet connection? It is likely that you've lost connection.");
-                        webSocket.close();
+                        webSocket.close(true);
+                    }
+                    catch (InterruptedException e)
+                    {
+                        //We've been asked to stop. The next iteration will kill the loop.
+                    }
+                    catch (SocketException e)
+                    {
+                        //Most likely the socket has been closed due to the audio connection be closed. Next iteration will kill loop.
                     }
                     catch (Exception e)
                     {
@@ -212,7 +219,15 @@ public class AudioConnection
             @Override
             public void run()
             {
-                while (!udpSocket.isClosed())
+                try
+                {
+                    udpSocket.setSoTimeout(100);
+                }
+                catch (SocketException e)
+                {
+                    LOG.log(e);
+                }
+                while (!udpSocket.isClosed() && !this.isInterrupted())
                 {
                     DatagramPacket receivedPacket = new DatagramPacket(new byte[1920], 1920);
                     try
@@ -229,6 +244,10 @@ public class AudioConnection
 
                             receiveHandler.handleReceivedAudio(decryptedPacket);
                         }
+                    }
+                    catch (SocketTimeoutException e)
+                    {
+                        //Ignore. We set a low timeout so that we wont block forever so we can properly shutdown the loop.
                     }
                     catch (SocketException e)
                     {
