+++ src/main/java/tomp2p/opuswrapper/Opus.java
@@ -8,7 +8,7 @@ import com.sun.jna.ptr.FloatByReference;
 import com.sun.jna.ptr.IntByReference;
 import com.sun.jna.ptr.PointerByReference;
 import com.sun.jna.ptr.ShortByReference;
-import net.dv8tion.jda.managers.AudioManager;
+import net.dv8tion.jda.managers.impl.AudioManagerImpl;
 
 import java.nio.ByteBuffer;
 import java.nio.FloatBuffer;
@@ -26,7 +26,7 @@ import java.nio.ShortBuffer;
  * Github repo.
  */
 public interface Opus extends Library {
-    public static final Opus INSTANCE = (Opus)Native.loadLibrary(AudioManager.OPUS_LIB_NAME, Opus.class);
+    public static final Opus INSTANCE = (Opus)Native.loadLibrary(AudioManagerImpl.OPUS_LIB_NAME, Opus.class);
 
     //  ****** Constants
